+++ src/Properties/MigrationHelper.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace NunoMaduro\Larastan\Properties;
 
-use Iterator;
 use PHPStan\File\FileHelper;
 use PHPStan\Parser\CachedParser;
 use RecursiveDirectoryIterator;
@@ -17,13 +16,16 @@ class MigrationHelper
     /** @var CachedParser */
     private $parser;
 
-    /** @var ?string */
+    /** @var string[] */
     private $databaseMigrationPath;
 
     /** @var string */
     private $currentWorkingDirectory;
 
-    public function __construct(CachedParser $parser, string $currentWorkingDirectory, ?string $databaseMigrationPath)
+    /**
+     * @param string[] $databaseMigrationPath
+     */
+    public function __construct(CachedParser $parser, string $currentWorkingDirectory, array $databaseMigrationPath)
     {
         $this->parser = $parser;
         $this->databaseMigrationPath = $databaseMigrationPath;
@@ -35,19 +37,17 @@ class MigrationHelper
      */
     public function initializeTables(): array
     {
-        if ($this->databaseMigrationPath !== null) {
-            $this->databaseMigrationPath = $this->getFileHelper()->absolutizePath($this->databaseMigrationPath);
-        } else {
-            $this->databaseMigrationPath = database_path('migrations');
-        }
-
-        if (! is_dir($this->databaseMigrationPath)) {
-            return [];
+        if (empty($this->databaseMigrationPath)) {
+            $this->databaseMigrationPath = [database_path('migrations')];
         }
 
         $schemaAggregator = new SchemaAggregator();
-        $files = $this->getMigrationFiles($this->databaseMigrationPath);
-        $filesArray = iterator_to_array($files);
+        $filesArray = $this->getMigrationFiles();
+
+        if (empty($filesArray)) {
+            return [];
+        }
+
         ksort($filesArray);
 
         $this->requireFiles($filesArray);
@@ -60,13 +60,25 @@ class MigrationHelper
     }
 
     /**
-     * @param string $path
-     *
-     * @return Iterator<SplFileInfo>
+     * @return SplFileInfo[]
      */
-    private function getMigrationFiles(string $path): Iterator
+    private function getMigrationFiles(): array
     {
-        return new RegexIterator(new RecursiveIteratorIterator(new RecursiveDirectoryIterator($path)), '/\.php$/i');
+        $migrationFiles = [];
+
+        foreach ($this->databaseMigrationPath as $additionalPath) {
+            $absolutePath = $this->getFileHelper()->absolutizePath($additionalPath);
+            if (is_dir($absolutePath)) {
+                $migrationFiles += iterator_to_array(
+                    new RegexIterator(
+                        new RecursiveIteratorIterator(new RecursiveDirectoryIterator($absolutePath)),
+                        '/\.php$/i'
+                    )
+                );
+            }
+        }
+
+        return $migrationFiles;
     }
 
     /**