+++ tests/Application/app/Group.php
@@ -17,4 +17,9 @@ class Group extends Model
     {
         return $this->hasMany(Account::class);
     }
+
+    public function users(): HasMany
+    {
+        return $this->hasMany(User::class);
+    }
 }