+++ tests/Features/Methods/Facades.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace Tests\Features\Methods;
 
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Event;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Queue;
 use Illuminate\Support\Facades\Redis;
 
 class Facades
@@ -47,5 +50,26 @@ class Facades
     public function testRedisHmget(): array
     {
         return Redis::hmget('h', ['field1', 'field2']);
+    }
+
+    public function testDBQuery(): \Illuminate\Database\Query\Builder
+    {
+        return DB::query();
+    }
+
+    public function testDBTransactionLevel(): int
+    {
+        return DB::transactionLevel();
+    }
+
+    public function testQueueCreatePayloadUsing(): void
+    {
+        Queue::createPayloadUsing(function () {
+        });
+    }
+
+    public function testGetLogger(): \Psr\Log\LoggerInterface
+    {
+        return Log::getLogger();
     }
 }