+++ tests/Unit/MigrationHelperTest.php
@@ -22,7 +22,7 @@ class MigrationHelperTest extends TestCase
     /** @test */
     public function it_will_return_empty_array_if_migrations_path_is_not_a_directory()
     {
-        $migrationHelper = new MigrationHelper($this->cachedParser, '', 'foobar');
+        $migrationHelper = new MigrationHelper($this->cachedParser, '', ['foobar']);
 
         self::assertSame([], $migrationHelper->initializeTables());
     }
@@ -30,7 +30,7 @@ class MigrationHelperTest extends TestCase
     /** @test */
     public function it_can_read_basic_migrations_and_create_table_structure()
     {
-        $migrationHelper = new MigrationHelper($this->cachedParser, '', __DIR__.'/data/basic_migration');
+        $migrationHelper = new MigrationHelper($this->cachedParser, '', [__DIR__.'/data/basic_migration']);
 
         $tables = $migrationHelper->initializeTables();
 
@@ -40,7 +40,7 @@ class MigrationHelperTest extends TestCase
     /** @test */
     public function it_can_read_schema_definitions_from_any_method_in_class()
     {
-        $migrationHelper = new MigrationHelper($this->cachedParser, '', __DIR__.'/data/migrations_with_different_methods');
+        $migrationHelper = new MigrationHelper($this->cachedParser, '', [__DIR__.'/data/migrations_with_different_methods']);
 
         $tables = $migrationHelper->initializeTables();
 
@@ -50,7 +50,7 @@ class MigrationHelperTest extends TestCase
     /** @test */
     public function it_can_read_schema_definitions_with_multiple_create_and_drop_methods_for_one_table()
     {
-        $migrationHelper = new MigrationHelper($this->cachedParser, '', __DIR__.'/data/complex_migrations');
+        $migrationHelper = new MigrationHelper($this->cachedParser, '', [__DIR__.'/data/complex_migrations']);
 
         $tables = $migrationHelper->initializeTables();
 
@@ -64,6 +64,21 @@ class MigrationHelperTest extends TestCase
         self::assertSame('string', $tables['users']->columns['created_at']->readableType);
         self::assertSame('string', $tables['users']->columns['updated_at']->readableType);
         self::assertSame('int', $tables['users']->columns['active']->readableType);
+    }
+
+    /** @test */
+    public function it_can_read_additional_directories(): void
+    {
+        $migrationHelper = new MigrationHelper($this->cachedParser, '', [
+            __DIR__.'/data/basic_migration',
+            __DIR__.'/data/additional_migrations',
+        ]);
+
+        $tables = $migrationHelper->initializeTables();
+
+        self::assertCount(2, $tables);
+        self::assertArrayHasKey('users', $tables);
+        self::assertArrayHasKey('teams', $tables);
     }
 
     /**