+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- BREAKING: Changed `databaseMigrationsPath` to be an array instead of a string, so it allows multiple directories.
+
 ### [0.6.13] - 2021-01-22
 
 ### Added