+++ CHANGELOG.md
@@ -6,7 +6,37 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-### [0.6.12] - 2021-01-03
+### Added
+- Support for HigherOrderCollectionProxy ([#781](https://github.com/nunomaduro/larastan/pull/781))
+
+## [0.7.0] - 2021-02-01
+### Changed
+- Updated `phpstan/phpstan` from v0.12.65 to v0.12.70 ([#774](https://github.com/nunomaduro/larastan/pull/774)) Thanks @chrisp-github
+
+### Added
+
+- **BREAKING**: Changed `databaseMigrationsPath` to be an array instead of a string, so it allows multiple directories. ([#777](https://github.com/nunomaduro/larastan/pull/777)) Thanks @haringsrob
+- Support for Log facade ([#771](https://github.com/nunomaduro/larastan/pull/771)) Thanks @Pe46dro
+
+## [0.6.13] - 2021-01-22
+
+### Added
+
+- DB facade stub ([8aa4b6d](https://github.com/nunomaduro/larastan/commit/8aa4b6dc3407fad6ced8115da02f610aad55366f))
+- Queue facade stub ([d0b0909](https://github.com/nunomaduro/larastan/commit/d0b0909c3ef60d45d3db9a2cf1fb91a73711ef26))
+- Logger stub ([208df29](https://github.com/nunomaduro/larastan/commit/208df29dab0ad80cabd6dc1b4e2232fc0f1044a0))
+
+### Changed
+
+- Use updated excludePaths parameter ([#759](https://github.com/nunomaduro/larastan/pull/759)) Thanks @erikgaal
+- Improve Pipes\Facades ([#744](https://github.com/nunomaduro/larastan/pull/744)) Thanks @szepeviktor
+
+### Fixed
+
+- Fixed insertOrIgnore typehint ([#756](https://github.com/nunomaduro/larastan/pull/756)) Thanks @bistory
+- Fix BelongsToMany stub's file name ([#736](https://github.com/nunomaduro/larastan/pull/736)) Thanks @szepeviktor
+
+## [0.6.12] - 2021-01-03
 
 ### Added
 
@@ -16,7 +46,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Search missing method also in templated model ([#739](https://github.com/nunomaduro/larastan/pull/739))
 
-### [0.6.11] - 2020-12-07
+## [0.6.11] - 2020-12-07
 
 ### Added
 
@@ -487,7 +517,10 @@ Upgrade guide: [UPGRADE.md](https://github.com/nunomaduro/larastan/blob/master/U
 ### Added
 - Adds first alpha version
 
-[Unreleased]: https://github.com/nunomaduro/larastan/compare/v0.6.11...HEAD
+[Unreleased]: https://github.com/nunomaduro/larastan/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/nunomaduro/larastan/compare/v0.6.13...v0.7.0
+[0.6.13]: https://github.com/nunomaduro/larastan/compare/v0.6.12...v0.6.13
+[0.6.12]: https://github.com/nunomaduro/larastan/compare/v0.6.11...v0.6.12
 [0.6.11]: https://github.com/nunomaduro/larastan/compare/v0.6.10...v0.6.11
 [0.6.10]: https://github.com/nunomaduro/larastan/compare/v0.6.9...v0.6.10
 [0.6.9]: https://github.com/nunomaduro/larastan/compare/v0.6.8...v0.6.9